requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
docopt==0.6.2
idna==3.10
pipreqs==0.4.13
psycopg2==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
yarg==0.1.10